+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 * Made y-axis markers consistent #124
 * Addressed OSX Data loss while running in background #139 using appnope package.
 * Fixed incorrect instantiation of Qt Thread from a python thread.  May fix #132.
+* Added "Widgets/Waveform/scale" property for default waveform scale #135
 
 
 ## 0.9.10